revert don't always shift left
