+++ std/compress/lzss/snark_testing.go
@@ -3,6 +3,7 @@ package lzss
 import (
 	"github.com/consensys/compress/lzss"
 	"github.com/consensys/gnark/frontend"
+	"github.com/consensys/gnark/std/compress"
 	test_vector_utils "github.com/consensys/gnark/std/utils/test_vectors_utils"
 )
 
@@ -21,7 +22,10 @@ type DecompressionTestCircuit struct {
 func (c *DecompressionTestCircuit) Define(api frontend.API) error {
 	dict := test_vector_utils.ToVariableSlice(lzss.AugmentDict(c.Dict))
 	dBack := make([]frontend.Variable, len(c.D)) // TODO Try smaller constants
-	dLen, err := Decompress(api, c.C, c.CBegin, c.CLength, dBack, dict, c.Level)
+	if cb, ok := c.CBegin.(int); !ok || cb != 0 {
+		c.C = compress.ShiftLeft(api, c.C, c.CBegin)
+	}
+	dLen, err := Decompress(api, c.C, c.CLength, dBack, dict, c.Level)
 	if err != nil {
 		return err
 	}
